requirements.txt
absl-py==2.0.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
Flask==3.0.0
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
idna==3.4
ipykernel==6.23.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.14.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
nest-asyncio==1.5.6
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.5.3
prompt-toolkit==3.0.38
protobuf==4.25.0
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pywin32==306
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.6.2
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.6
Werkzeug==3.0.1
wrapt==1.14.1
